bug fix- wrong cast while generating new PassCode

+++ Bills/app/src/main/java/com/bills/bills/BillsMainActivity.java
@@ -483,7 +483,7 @@ public class BillsMainActivity extends MainActivityBase implements IOnCameraFini
                     mutableData.child(mUid).setValue(value);
                 } else {
                         for (MutableData childMutableData : mutableData.getChildren()) {
-                            passCodes.put(childMutableData.getKey(), (Integer)((HashMap<String, Object>)childMutableData.getValue()).get(mPassCodeDbKey));
+                            passCodes.put(childMutableData.getKey(), ((Long)(((HashMap<String, Object>)childMutableData.getValue()).get(mPassCodeDbKey))).intValue());
                         }
                     //find an unused pass code
                     for (int i = 0; i < 10000; i++) {
